refactor(wash-cli): refactor link management commands
This commit moves the link management commands (get, put, query) into
the new cmd/ structure.
