+++ crates/wash-cli/src/cmd/mod.rs
@@ -2,5 +2,6 @@
 
 pub mod config;
 pub mod dev;
+pub mod link;
 pub mod up;
 pub mod wit;
